refactor support page and add support form component
