+++ src/app/(content)/layout.tsx
@@ -6,10 +6,9 @@ export default function Layout({
   children: React.ReactNode;
 }>) {
   return (
+    // change to correct layout, which is herader and footer with small logo
     <AdditionalNavigationLayout>
-      <div className="px-4 text-sm md:text-base lg:px-32 lg:py-10 lg:text-lg">
-        {children}
-      </div>
+      <div className="px-4 text-sm lg:px-32 lg:py-10">{children}</div>
     </AdditionalNavigationLayout>
   );
 }
